requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
celery==5.3.6
loguru==0.7.2
tenacity==8.2.3
feedparser==6.0.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
transformers==4.36.2
torch==2.1.2
spacy==3.7.2
scikit-learn==1.3.2
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.20.0
psutil==5.9.8
email-validator==2.1.0.post1
stripe==11.6.0
